+++ src/app/dataset-view/additional-components/metadata-component/metadata.component.ts
@@ -44,6 +44,9 @@ export class MetadataComponent extends BaseComponent implements OnInit {
         ReadStepGeoJson: "Geo json",
         ReadStepEsriShapefile: "Esri shapefile",
         ReadStepParquet: "Parquet",
+        ReadStepJson: "Json",
+        ReadStepNdJson: "Newline-delimited json",
+        ReadStepNdGeoJson: "Newline-delimited geo json",
     };
 
     public currentState?: {